Move FitMode to configuration.h
