+++ YACReader/configuration.h
@@ -1,5 +1,6 @@
 #ifndef __CONFIGURATION_H
 #define __CONFIGURATION_H
+
 #include <QByteArray>
 #include <QString>
 #include <QSize>
@@ -14,6 +15,16 @@
 #define SLIDE_ASPECT_RATIO 1.585
 
 using namespace YACReader;
+
+namespace YACReader {
+
+enum FitMode {
+    ToWidth = 0x01,
+    ToHeight = 0x02,
+    FullRes = 0x03,
+    FullPage = 0x04 //,
+    // Text=0x05
+};
 
 class Configuration : public QObject
 {
@@ -85,6 +96,7 @@ public:
     bool getUseSingleScrollStepToTurnPage() { return settings->value(USE_SINGLE_SCROLL_STEP_TO_TURN_PAGE, false).toBool(); }
     void setDisableScrollAnimation(bool b) { settings->setValue(DISABLE_SCROLL_ANIMATION, b); }
     bool getDisableScrollAnimation()
+
     {
 #ifdef Q_OS_MACOS
         auto defaultValue = true;
@@ -95,5 +107,7 @@ public:
         return settings->value(DISABLE_SCROLL_ANIMATION, defaultValue).toBool();
     }
 };
+
+}
 
 #endif
